Consolidate bridge file-store plumbing into store_util
- Move the byte-identical private store toolkit duplicated across
  notes.rs and agent_pins.rs (session_key, stable_hash, now_ms_string,
  env/XDG dir helpers, private dir/file permissions, corrupt-copy
  handling, flock LockFile) into a shared bridge/src/store_util.rs.
- Merge run_notes_task and run_agent_pins_task (identical generic
  spawn_blocking wrappers) into one run_store_task.
- Switch agent_activity.rs to the shared session_key/stable_hash so
  /api/agent-activity reports session keys consistent with notes and
  agent-pins (state is memory-only; only the reported string format
  changes).
- Keep write_json_atomic per-store (the notes variant writes .bak
  backups) and keep the upload-dir env helper local to web_bridge.rs
  (it expands ~, the store variant intentionally does not).
- Add .truncate(false) to the flock-only LockFile open to document
  intent and clear the clippy suspicious_open_options warning.

Net: ~350 lines removed.
