+++ bridge/src/main.rs
@@ -2,6 +2,7 @@ mod agent_activity;
 mod agent_pins;
 mod notes;
 mod session;
+mod store_util;
 mod web_bridge;
 mod workspace;
